use typescript and add image upload feature
