+++ src/action/index.js
@@ -1,0 +1,8 @@
+import { ActionType } from "../reducers/content";
+
+export const fillContent = () => {
+  return {
+    type: ActionType.FILL_CONTENT,
+    payload: { content },
+  };
+};
