--- src/contexts/ContentContext.js
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const ContentContext = createContext();
-
-export default ContentContext;
